+++ db-backup/restore.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+psql -U et -W -h localhost qa < $1