+++ CHANGELOG.md
@@ -7,11 +7,37 @@
 - docker? docker-compose? docker してた方が開発が楽か？
 - イメージファイルをアップロードできる。
 - 短すぎる質問をリジェクト。
-- 「いいね」の集計ページを設ける
+- 「いいね」の集計ページ
+- find-by-key, :order-by option
+
+## 0.4.3 - 2021-07-08
+### Changed
+- ニックネームをクリックで QA ページへ。
+
+## 0.4.2 - 2021-07-08
+### Added
+- unwscape-br answer でのみ、<br> 復活させる。
+
+## 0.4.1 - 2021-07-08
+### Added
+- /admin, /admin-goods 誰が good したかを表示する。
+
+## 0.4.0 - 2021-07-05
+- [:a {:href "/qs" :class "btn btn-success btn-sm"} "questions"]]
+- /qs 回答の表示順。jdbc/query でかけるのだが、jdbc/find-by-keys で書けない。
+
+
+## 0.3.4 - 2021-06-28
+### Changed
+- /qs の並びを新着順に変更。
+
+## 0.3.3 - 2021-06-23
+- オープン戦終了。page.clj から必要のない li を削除。
 
 ## 0.3.2 - 2021-06-19
 - systemd
-- 全角スペースを page.clj から剥ぎ取る。時々、zenkaku は全角スペースを表示しない。
+- 全角スペースを page.clj から剥ぎ取る。
+  時々、zenkaku は全角スペースを表示しない。
 
 ## 0.3.1 - 2021-06-19
 - ログインにバグ。ニックネームが "" の学生がある。
@@ -47,7 +73,7 @@
 - /as/:n エンドポイント。
 
 ## 0.1.1 - 2021-06-18
-- middleware　auth をセット。
+- middleware auth をセット。
 - initdb.d/{up,down}.bb で initdb.d/*.sql をリプレース。
 - ログインできる。
 - ページのボトムに logout ボタン。